+++ source/others/extF80M_to_f16.c
@@ -34,28 +34,26 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 =============================================================================*/
 
-#include <stdbool.h>
-#include <stdint.h>
+#include "softfloat/functions.h"
 
 #include "internals.h"
 #include "specialize.h"
-#include "softfloat/functions.h"
 
 #ifdef SOFTFLOAT_FAST_INT64
 
-float16_t extF80M_to_f16( const extFloat80_t *aPtr )
+float16_t extF80M_to_f16(const extFloat80_t *aPtr)
 {
 
-    return extF80_to_f16( *aPtr );
+    return extF80_to_f16(*aPtr);
 
 }
 
 #else
 
-float16_t extF80M_to_f16( const extFloat80_t *aPtr )
+float16_t extF80M_to_f16(const extFloat80_t *aPtr)
 {
     const struct extFloat80M *aSPtr;
-    uint_fast16_t uiA64;
+    uint16_t uiA64;
     bool sign;
     int32_t exp;
     uint64_t sig;
@@ -70,40 +68,40 @@ float16_t extF80M_to_f16( const extFloat80_t *aPtr )
     /*------------------------------------------------------------------------
     *------------------------------------------------------------------------*/
     uiA64 = aSPtr->signExp;
-    sign = signExtF80UI64( uiA64 );
-    exp  = expExtF80UI64( uiA64 );
+    sign = signExtF80UI64(uiA64);
+    exp = expExtF80UI64(uiA64);
     sig = aSPtr->signif;
     /*------------------------------------------------------------------------
     *------------------------------------------------------------------------*/
-    if ( exp == 0x7FFF ) {
-        if ( sig & UINT64_C( 0x7FFFFFFFFFFFFFFF ) ) {
-            softfloat_extF80MToCommonNaN( aSPtr, &commonNaN );
-            uiZ = softfloat_commonNaNToF16UI( &commonNaN );
+    if (exp == 0x7FFF) {
+        if (sig & UINT64_C(0x7FFFFFFFFFFFFFFF)) {
+            softfloat_extF80MToCommonNaN(aSPtr, &commonNaN);
+            uiZ = softfloat_commonNaNToF16UI(&commonNaN);
         } else {
-            uiZ = packToF16UI( sign, 0x1F, 0 );
+            uiZ = packToF16UI(sign, 0x1F, 0);
         }
         goto uiZ;
     }
     /*------------------------------------------------------------------------
     *------------------------------------------------------------------------*/
-    if ( ! (sig & UINT64_C( 0x8000000000000000 )) ) {
-        if ( ! sig ) {
-            uiZ = packToF16UI( sign, 0, 0 );
+    if (!(sig & UINT64_C(0x8000000000000000))) {
+        if (!sig) {
+            uiZ = packToF16UI(sign, 0, 0);
             goto uiZ;
         }
-        exp += softfloat_normExtF80SigM( &sig );
+        exp += softfloat_normExtF80SigM(&sig);
     }
     /*------------------------------------------------------------------------
     *------------------------------------------------------------------------*/
-    sig16 = softfloat_shortShiftRightJam64( sig, 49 );
+    sig16 = (uint16_t)softfloat_shortShiftRightJam64(sig, 49);
     exp -= 0x3FF1;
-    if ( sizeof (int_fast16_t) < sizeof exp ) {
-        if ( exp < -0x40 ) exp = -0x40;
+    if (sizeof(int16_t) < sizeof exp) {
+        if (exp < -0x40) exp = -0x40;
     }
-    return softfloat_roundPackToF16( sign, exp, sig16 );
+    return softfloat_roundPackToF16(sign, exp, sig16);
     /*------------------------------------------------------------------------
     *------------------------------------------------------------------------*/
- uiZ:
+uiZ:
     uZ.ui = uiZ;
     return uZ.f;
